Refactor the packages to separate the API server functionality. Change the pom to generate fat jar
