+++ src/saasy-ml-app/pom.xml
@@ -90,7 +90,7 @@
       <version>3.12.0</version>
     </dependency>
 
-    <!-- Vertx Dependencies -->
+    <!-- API server Dependencies -->
     <dependency>
       <groupId>io.vertx</groupId>
       <artifactId>vertx-config</artifactId>
@@ -150,6 +150,29 @@
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-resources-plugin</artifactId>
       </plugin>
+            
+      <plugin>
+          <artifactId>maven-assembly-plugin</artifactId>
+          <configuration>
+              <archive>
+                  <manifest>
+                      <mainClass>${assembly.mainClass}</mainClass>
+                  </manifest>
+              </archive>
+              <descriptorRefs>
+                  <descriptorRef>jar-with-dependencies</descriptorRef>
+              </descriptorRefs>
+          </configuration>
+          <executions>
+              <execution>
+                  <id>make-assembly</id>
+                  <phase>package</phase>
+                  <goals>
+                      <goal>single</goal>
+                  </goals>
+              </execution>
+          </executions>
+      </plugin>
     </plugins>
   </build>
   
@@ -173,6 +196,7 @@
                 </execution>
               </executions>
             </plugin>
+
         </plugins>
       </build>
     </profile>
