+++ src/saasy-ml-app/src/main/java/esa/mo/nmf/apps/saasyml/api/exceptions/AddAggregationDidNotReturnAggregationId.java
@@ -1,4 +1,4 @@
-package esa.mo.nmf.apps.exceptions;
+package esa.mo.nmf.apps.saasyml.api.exceptions;
 
 public class AddAggregationDidNotReturnAggregationId extends Exception {
     private static final long serialVersionUID = 1L;
