+++ src/saasy-ml-app/src/main/java/esa/mo/nmf/apps/saasyml/api/exceptions/NotEnoughParamsForRequestedType.java
@@ -1,4 +1,4 @@
-package esa.mo.nmf.apps.exceptions;
+package esa.mo.nmf.apps.saasyml.api.exceptions;
 
 public class NotEnoughParamsForRequestedType extends Exception {
     private static final long serialVersionUID = 1L;
